fix: refactor cloudant db connection
