+++ src/database/index.js
@@ -1,3 +1,50 @@
-import CloudantDatabase from "./cloudant";
+import Cloudant from "@cloudant/cloudant";
+import { cloudant as config } from "../config";
+import createLogger from "../utils/logger";
 
-export default CloudantDatabase;
+const logger = createLogger({
+  name: "cloudant-logger",
+});
+
+const cloudantConfig = {
+  account: config.cloudantId,
+  plugins: {
+    iamauth: {
+      iamApiKey: config.apiKey,
+    },
+  },
+};
+
+export const connectCloudantDatabase =  () => {
+    return Cloudant(cloudantConfig, (err, cloudant) => {
+      if (err) {
+        logger.error(
+          `Connect failure: ${err.message} for Cloudant ID: ${config.cloudantId}`
+        );
+        return error;
+      } else {
+        (async () => {
+          try {
+            const existingDbs = await cloudant.db.list();
+            if (!existingDbs.includes(config.dbName)) {
+              logger.info(
+                `Database does not exist... creating: ${config.dbName}`
+              );
+              await cloudant.db.create(config.dbName);
+            }
+            const db = cloudant.use(config.dbName);
+            logger.info(
+              `Connection success! Connected to DB: ${config.dbName}`
+            );
+            return db;
+          } catch (error) {
+            logger.error(
+              `Connect failure: ${err.message} for Cloudant ID: ${config.cloudantId}`
+            );
+            return error;
+          }
+        })();
+      }
+    });
+}
+
