+++ src/index.js
@@ -1,10 +1,12 @@
 import { ApolloServer, PubSub } from "apollo-server";
 import modules from "./modules";
 import plugins from "./plugins";
-import CloudantDatabase from "./database";
+import { connectCloudantDatabase } from "./database";
+import  CloudantDatabase from "./database/cloudant";
 
-const database = new CloudantDatabase();
 
+//connectCloudantDatabase connects to the cloudant db and returns thhe database
+const database = new CloudantDatabase(connectCloudantDatabase());
 // Anything shared between resolvers are put into context
 const server = new ApolloServer({
   plugins,
